+++ packages/patternfly-4/react-charts/src/components/ChartDonutUtilization/index.js
@@ -1,0 +1,2 @@
+export { default as ChartDonutThreshold } from './ChartDonutThreshold';
+export { default as ChartDonutUtilization } from './ChartDonutUtilization';