+++ packages/patternfly-4/react-charts/src/components/ChartLegend/ChartLegend.js
@@ -5,7 +5,7 @@ import hoistNonReactStatics from 'hoist-non-react-statics';
 import { VictoryLegend } from 'victory';
 import ChartContainer from '../ChartContainer/ChartContainer';
 import ChartPoint from '../ChartPoint/ChartPoint';
-import { getTheme } from '../ChartTheme/themes/theme-utils';
+import { getTheme } from '../ChartUtils/chart-theme';
 
 export const propTypes = {
   /**
@@ -13,7 +13,7 @@ export const propTypes = {
    */
   '': PropTypes.any,
   /**
-   * The height props specifies the height the svg viewBox of the chart container. This value should be given as a
+   * Specifies the height the svg viewBox of the chart container. This value should be given as a
    * number of pixels.
    *
    * Because Victory renders responsive containers, the width and height props do not determine the width and
@@ -27,17 +27,25 @@ export const propTypes = {
    *
    * Useful when legend is located inside a chart -- default is false.
    */
-  responsive: PropTypes.boolean,
-  /*
-   * Specifies the theme color; blue (default), green, or multi-color. Overridden by the theme property.
+  responsive: PropTypes.bool,
+  /**
+   * Specifies the theme color. Valid values are 'blue', 'green', 'grey' (recomended), 'multi', etc.
+   *
+   * Note: Not compatible with theme prop
+   *
+   * @example themeColor={ChartThemeColor.blue}
    */
   themeColor: PropTypes.string,
-  /*
-   * Specifies the theme variant; 'dark' or 'light' (default). Overridden by the theme property.
+  /**
+   * Specifies the theme variant. Valid values are 'dark' or 'light'
+   *
+   * Note: Not compatible with theme prop
+   *
+   * @example themeVariant={ChartThemeVariant.light}
    */
   themeVariant: PropTypes.string,
   /**
-   * The width props specifies the width of the svg viewBox of the chart container. This value should be given as a
+   * Specifies the width of the svg viewBox of the chart container. This value should be given as a
    * number of pixels.
    *
    * Because Victory renders responsive containers, the width and height props do not determine the width and
@@ -48,7 +56,7 @@ export const propTypes = {
 };
 
 const ChartLegend = ({
-  responsive = false,
+  responsive = true,
   themeColor,
   themeVariant,
   theme = getTheme(themeColor, themeVariant), // destructure last