+++ packages/patternfly-3/patternfly-react-extensions/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## 2.18.11 (2019-05-29)
+
+**Note:** Version bump only for package patternfly-react-extensions
+
+
+
+
+
 ## 2.18.10 (2019-05-20)
 
 **Note:** Version bump only for package patternfly-react-extensions