+++ packages/patternfly-4/react-inline-edit-extension/CHANGELOG.md
@@ -3,6 +3,17 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.4.15](https://github.com/patternfly/patternfly-react/compare/@patternfly/react-inline-edit-extension@2.4.14...@patternfly/react-inline-edit-extension@2.4.15) (2019-05-29)
+
+
+### Bug Fixes
+
+* **docs:** use frontmatter for what props to show in docs ([#2072](https://github.com/patternfly/patternfly-react/issues/2072)) ([20e6b12](https://github.com/patternfly/patternfly-react/commit/20e6b12))
+
+
+
+
+
 ## [2.4.14](https://github.com/patternfly/patternfly-react/compare/@patternfly/react-inline-edit-extension@2.4.13...@patternfly/react-inline-edit-extension@2.4.14) (2019-05-28)
 
 **Note:** Version bump only for package @patternfly/react-inline-edit-extension